EquipmentManagement/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.1.0
Flask-Migrate==4.1.0
Flask-WTF==1.2.1